+++ examples/account-public-access/outputs.tf
@@ -1,0 +1,4 @@
+output "s3_account_public_access_block_id" {
+  description = "AWS account ID"
+  value       = module.account_public_access.s3_account_public_access_block_id
+}