+++ modules/account-public-access/versions.tf
@@ -4,11 +4,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 5.83"
-    }
-    random = {
-      source  = "hashicorp/random"
-      version = ">= 2.0"
+      version = ">= 3.74"
     }
   }
 }