+++ modules/notification/main.tf
@@ -60,7 +60,7 @@ resource "aws_s3_bucket_notification" "this" {
 
 # Lambda
 resource "aws_lambda_permission" "allow" {
-  for_each = var.lambda_notifications
+  for_each = { for k, v in var.lambda_notifications : k => v if var.create_lambda_permission }
 
   statement_id_prefix = "AllowLambdaS3BucketNotification-"
   action              = "lambda:InvokeFunction"
@@ -80,6 +80,8 @@ data "aws_arn" "queue" {
 
 data "aws_iam_policy_document" "sqs" {
   for_each = { for k, v in var.sqs_notifications : k => v if var.create_sqs_policy }
+
+  version = "2012-10-17"
 
   statement {
     sid = "AllowSQSS3BucketNotification"